associateChanges
