+++ src/pages/Admin/AdminLogin/style.scss
@@ -4,7 +4,7 @@
     display: flex;
     align-items: center;
     justify-content: center;
-    background-image: url('../../assets/admin-bg.jpg');
+    background-image: url('../../../assets/admin-bg.jpg');
     background-position: center;
     background-repeat: no-repeat;
     background-size: cover;
