+++ src/pages/Associates/Layout/style.scss
