+++ src/pages/Layout/Layout.js
@@ -11,26 +11,16 @@ import Divider from "@mui/material/Divider";
 import MenuIcon from "@mui/icons-material/Menu";
 import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
 import ChevronRightIcon from "@mui/icons-material/ChevronRight";
-import ListItem from "@mui/material/ListItem";
-import ListItemButton from "@mui/material/ListItemButton";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import ListItemText from "@mui/material/ListItemText";
+
 import Avatar from "@mui/material/Avatar";
-import ChatIcon from "@mui/icons-material/Chat";
-import FavoriteIcon from "@mui/icons-material/Favorite";
-import ManageAccountsIcon from "@mui/icons-material/ManageAccounts";
-import DisplaySettingsIcon from "@mui/icons-material/DisplaySettings";
-import CompareArrowsIcon from "@mui/icons-material/CompareArrows";
-import PreviewIcon from "@mui/icons-material/Preview";
-import HomeIcon from "@mui/icons-material/Home";
 import IconButton from "@mui/material/IconButton";
 import { NavLink, Outlet, useNavigate, useLocation } from "react-router-dom";
 import Menu from "@mui/material/Menu";
 import Tooltip from "@mui/material/Tooltip";
 import MenuItem from "@mui/material/MenuItem";
 import Authorization from "../../utils/authorization";
-import "./style.scss";
-import { useState, useEffect } from "react";
+import AssociateLayout from "../Associates/Layout/Layout";
+import UserLayout from "../Users/Layout/Layout";
 
 const drawerWidth = 280;
 
@@ -138,8 +128,6 @@ const Layout = () => {
   const location = useLocation();
   const isAssociatelogin = location && location.pathname && location.pathname.includes('associates');
 
-
-
   const handleClick = (event) => {
     setAnchorEl(event.currentTarget);
   };
@@ -157,68 +145,6 @@ const Layout = () => {
     navigate("associates/viewprofile");
   };
 
-  let sideMenu = [
-    {
-      title: "Home",
-      icon: "https://cdn.lordicon.com/gmzxduhd.json",
-      path: "home",
-      role: "USER",
-    },
-    {
-      title: "Wish List",
-      icon: "https://cdn.lordicon.com/rjzlnunf.json",
-      path: "wishlist",
-      role: "USER",
-    },
-    {
-      title: "Chat",
-      icon: "https://cdn.lordicon.com/zpxybbhl.json",
-      path: "chat",
-      role: "USER",
-    },
-    {
-      title: "Edit Profile",
-      icon: "https://cdn.lordicon.com/wloilxuq.json",
-      path: "edit-profile",
-      role: "USER",
-    },
-    {
-      title: "Edit Preference",
-      icon: "https://cdn.lordicon.com/sbiheqdr.json",
-      path: "edit-preference",
-      role: "USER",
-    },
-    {
-      title: "Compare Profiles",
-      icon: "https://cdn.lordicon.com/soseozvi.json",
-      path: "compare-profiles",
-      role: "USER",
-    },
-    {
-      title: "Profile Viewed",
-      icon: "https://cdn.lordicon.com/tyounuzx.json",
-      path: "profile-viewed",
-      role: "USER",
-    },
-    {
-      title: "Admin Dashboard",
-      icon: "https://cdn.lordicon.com/imamsnbq.json",
-      path: "admin-dashboard",
-      role: "SUPER_ADMIN",
-    },
-    {
-      title: "Sub-Ordinates",
-      icon: "https://cdn.lordicon.com/eszyyflr.json",
-      path: "sub-ordinates",
-      role: "SUPER_ADMIN",
-    },
-    {
-      title: "Associates",
-      icon: "https://cdn.lordicon.com/imamsnbq.json",
-      path: "associates",
-      role: "SUPER_ADMIN",
-    },
-  ];
 
   const handleDrawerOpen = () => {
     setOpen(true);
@@ -228,28 +154,6 @@ const Layout = () => {
     setOpen(false);
   };
 
-  if (isAssociatelogin) {
-    sideMenu = [
-      {
-        title: "View Profile",
-        icon: "https://cdn.lordicon.com/imamsnbq.json",
-        path: "associates/viewprofile",
-        role: "USER",
-      },
-      {
-        title: "Edit Profile",
-        icon: "https://cdn.lordicon.com/wloilxuq.json",
-        path: "associates/editprofile",
-        role: "USER",
-      },
-      {
-        title: "Earnings",
-        icon: "https://cdn.lordicon.com/zpxybbhl.json",
-        path: "associates/Earnings",
-        role: "USER",
-      },
-    ];
-  }
   return (
     <Box>
       <CssBaseline />
@@ -305,51 +209,12 @@ const Layout = () => {
           </IconButton>
         </DrawerHeader>
         <Divider />
-        <List sx={{ paddingTop: 0 }}>
-          {sideMenu.map((menu, index) =>
-            menu.role === "USER" ? (
-              <NavLink
-                to={menu.path}
-                className={({ isActive }) =>
-                  isActive ? "activeRoute" : "routes"
-                }
-                key={index}
-              >
-                <ListItem
-                  key={menu.title}
-                  disablePadding
-                  sx={{ display: "block" }}
-                >
-                  <ListItemButton
-                    sx={{
-                      minHeight: 48,
-                      justifyContent: open ? "initial" : "center",
-                      padding: 0,
-                    }}
-                  >
-                    <Tooltip title={menu.title} placement="right">
-                      <lord-icon
-                        src={menu.icon}
-                        trigger="loop"
-                        colors="primary:#121331,secondary:#d53833"
-                        style={{
-                          width: open ? "60px" : "500px",
-                          marginLeft: open ? "auto" : "10px",
-                        }}
-                      ></lord-icon>
-                    </Tooltip>
 
-                    <ListItemText
-                      className="menu-text"
-                      primary={menu.title}
-                      sx={{ opacity: open ? 1 : 0 }}
-                    />
-                  </ListItemButton>
-                </ListItem>
-              </NavLink>
-            ) : null
-          )}
-        </List>
+        {isAssociatelogin ?
+          <AssociateLayout open={open}/> :
+          <UserLayout open={open}/>
+        }
+
       </Drawer>
       <Main open={open}>
         <DrawerHeader />
