+++ src/pages/Users/Profile/Profile.js
@@ -14,13 +14,13 @@ import {
 import { ToastContainer, toast, Zoom } from "react-toastify";
 import { useTheme } from "@mui/material/styles";
 import { useEffect, useState } from "react";
-import { getProfileData } from "../../api/api";
+import { getProfileData } from "../../../api/api";
 import EditIcon from "@mui/icons-material/Edit";
 import SwipeableViews from "react-swipeable-views";
 import { loadAnimation } from "lottie-web";
 import { defineLordIconElement } from "lord-icon-element";
 import moment from "moment";
-import { ls } from "../../utils/localStorage";
+import { ls } from "../../../utils/localStorage";
 
 // register lottie and define custom element
 defineLordIconElement(loadAnimation);
