+++ src/pages/Users/Profile/style.scss
@@ -3,7 +3,7 @@
         width: 350px;
         height: 350px;
         border-radius: 10px;
-        background-image: url('../../assets/img/testimonials/testimonials-5.jpg');
+        background-image: url('../../../assets/img/testimonials/testimonials-5.jpg');
         background-position: center;
         background-size: cover;
     }
