+++ src/pages/Users/ViewedProfile/ViewedProfile.scss
